+++ test/index.js
@@ -1,25 +1,25 @@
-'use strict';
-import assert from 'assert';
-import latex from '../lib/latex';
-import React from 'react';
+import React from "react";
+import assert from "assert";
+import ReactDomServer from "react-dom/server";
+import latex from "./../lib/Latex";
 
-var Test = React.createClass({
-    render: function() {
+class Test extends React.Component {
+    render() {
         return (
             React.createElement(latex, null,
                "What is $\\sqrt{8}$?"
             )
         );
     }
+}
 
-});
 
+describe("react-latex", () => {
+    it("Should have katex class", () => {
+        const testString = ReactDomServer.renderToStaticMarkup(<Test />);
 
-describe('react-latex', function () {
-    it('Should have katex class', function () {
-        var testString = React.renderToStaticMarkup(<Test/>);
-
-        assert.notEqual(-1, testString.indexOf('<span class="katex-mathml">'));
-
+        assert.notEqual(-1, testString.indexOf("<span class=\"katex-mathml\">"));
     });
+
+    it("Should ")
 });