+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 CHANGELOG
 =========
 
+# 1.0.0
+- Katex updated to 0.7.1
+- Minimum version of React bumped to 15.3.0
+- Fixed `displayMode` not being sent properly to Katex
+
 # 0.1.1
 - Fix gulp nsp
 
 # 0.1.0
-- Added support for Katex displayMode
+- Added support for Katex displayMode (Thanks to @mathisonian)